requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-aws>=0.1.0
langchain-community>=0.2.0
langsmith>=0.1.0
boto3>=1.34.0
sagemaker>=2.200.0
pinecone>=5.0.0
databricks-sdk>=0.20.0
pyarrow>=15.0.0
pandas>=2.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0